+++ CHANGELOG.md
@@ -15,6 +15,10 @@ All changes are categorized into one of the following keywords:
 
 
 ## 0.20.3 - 2012/01/24
+- **BUG** floatingmenu: Fixed positioning of floating menu when it extends
+          beyond the width of the viewport.
+- **BUG** floatingmenu: Fixed float position of floatingmenu when it moves
+          between editables.
 - **BUG** core: Removes ExtJS' IE6 style fixes which break layout in IE9.
 - **BUG** image-plugin: The image plugin will now only display the crop buttons
           when the cropping area selection was finished. This avoids a bug in